refactor: unique device identifier
